requirements.txt
streamlit>=1.36,<2.0
pandas>=2.0,<3.0
numpy>=1.24,<3.0
PyYAML>=6.0,<7.0
streamlit-autorefresh>=1.0.1,<2.0
python-dotenv>=1.0,<2.0
tushare>=1.4.24
openai>=1.0,<2.0
